server/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==3.2.7
django-cors-headers==3.8.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
Pillow==9.0.1
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
